requirements.txt
fastapi==0.88.0
pytest==7.1.3
pytest-asyncio==0.20.3
uvicorn==0.20.0
pydantic~=1.10.4
line-bot-sdk==1.8.0
openai==1.54.1
langchain-core>=0.2.43,<0.3
langchain>=0.2,<0.3
langchain-openai
langchain_community>=0.2,<0.3
streamlit
libsass